requirements.txt
# UI Framework
PyQt6==6.6.1

# Image Processing (Pinned to avoid NumPy 2.0 conflicts)
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0

# File Parsers
psd-tools==1.9.33
watchdog==4.0.0
